Changed file organization and directories and refactored getters code
