+++ tabs/bomlist/bomlist.php
@@ -4,7 +4,9 @@
 
 // Incluir configuração da base de dados
 // Incluir arquivo de configuração
-include_once __DIR__ . '/../PWA/RestAPI/config.php';
+include_once __DIR__ . '/../../PWA/RestAPI/config.php';
+
+require_once 'getters.php';
 
 // Estabelecer conexão com a base de dados
 $sqlFile = __DIR__ . '/database/database.sql';
@@ -27,121 +29,13 @@ try {
 }
 
 
-// try {
-//     $pdo = new PDO("mysql:host=$db_host;dbname=$db_name;charset=utf8mb4", $db_user, $db_pass);
-//     $pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-// } catch(PDOException $e) {
-//     die("Erro de conexão: " . $e->getMessage() .$db_host ) ;
-// }
-
-// Criar tabelas se não existirem
-// $createTablesSQL = [
-//     "CREATE TABLE IF NOT EXISTS T_Manufacturer (
-//         Manufacturer_ID INT AUTO_INCREMENT PRIMARY KEY,
-//         Denomination VARCHAR(255) NOT NULL,
-//         Origin_Country VARCHAR(100),
-//         Website VARCHAR(255),
-//         Contacts TEXT,
-//         Created_Date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-//         Updated_Date TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
-//     )",
-//     "CREATE TABLE IF NOT EXISTS T_Supplier (
-//         Supplier_ID INT AUTO_INCREMENT PRIMARY KEY,
-//         Denomination VARCHAR(255) NOT NULL,
-//         Origin_Country VARCHAR(100),
-//         Website VARCHAR(255),
-//         Contacts TEXT,
-//         Created_Date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-//         Updated_Date TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
-//     )",
-//     "CREATE TABLE IF NOT EXISTS T_Prototype (
-//         Prototype_ID INT AUTO_INCREMENT PRIMARY KEY,
-//         Name VARCHAR(255) NOT NULL,
-//         Version VARCHAR(50) DEFAULT '1.0',
-//         Description TEXT,
-//         Status ENUM('Development', 'Testing', 'Production', 'Archived') DEFAULT 'Development',
-//         Created_Date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-//         Updated_Date TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
-//     )",
-//     "CREATE TABLE IF NOT EXISTS T_Component (
-//         Component_ID INT AUTO_INCREMENT PRIMARY KEY,
-//         Denomination VARCHAR(255) NOT NULL,
-//         Manufacturer_ID INT,
-//         Manufacturer_ref VARCHAR(100),
-//         Supplier_ID INT,
-//         Supplier_ref VARCHAR(100),
-//         General_Type VARCHAR(100),
-//         Price DECIMAL(10,2),
-//         Acquisition_Date DATE,
-//         Notes_Description TEXT,
-//         Stock_Quantity INT DEFAULT 0,
-//         Min_Stock INT DEFAULT 0,
-//         Created_Date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-//         Updated_Date TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-//         FOREIGN KEY (Manufacturer_ID) REFERENCES T_Manufacturer(Manufacturer_ID) ON DELETE SET NULL,
-//         FOREIGN KEY (Supplier_ID) REFERENCES T_Supplier(Supplier_ID) ON DELETE SET NULL
-//     )",
-//     "CREATE TABLE IF NOT EXISTS T_Assembly (
-//         Assembly_ID INT AUTO_INCREMENT PRIMARY KEY,
-//         Prototype_ID INT NOT NULL,
-//         Father_ID INT,
-//         Child_ID INT NOT NULL,
-//         Quantity INT NOT NULL DEFAULT 1,
-//         Level_Depth INT DEFAULT 0,
-//         Notes TEXT,
-//         Created_Date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-//         FOREIGN KEY (Prototype_ID) REFERENCES T_Prototype(Prototype_ID) ON DELETE CASCADE,
-//         FOREIGN KEY (Father_ID) REFERENCES T_Component(Component_ID) ON DELETE CASCADE,
-//         FOREIGN KEY (Child_ID) REFERENCES T_Component(Component_ID) ON DELETE CASCADE,
-//         UNIQUE KEY unique_assembly (Prototype_ID, Father_ID, Child_ID)
-//     )"
-// ];
-
-// // Executar criação das tabelas individualmente
-// try {
-//     foreach ($createTablesSQL as $sql) {
-//         $pdo->exec($sql);
-//     }
-// } catch(PDOException $e) {
-//     echo "<div class='alert alert-warning'>Aviso ao criar tabelas: " . $e->getMessage() . "</div>";
-// }
 
 // Processar ações CRUD
 $action = $_POST['action'] ?? $_GET['action'] ?? 'list';
 $entity = $_POST['entity'] ?? $_GET['entity'] ?? 'components';
 $message = '';
 
-// Função para buscar todos os fabricantes
-function getManufacturers($pdo) {
-    $stmt = $pdo->query("SELECT * FROM T_Manufacturer ORDER BY Denomination");
-    return $stmt->fetchAll(PDO::FETCH_ASSOC);
-}
 
-// Função para buscar todos os fornecedores
-function getSuppliers($pdo) {
-    $stmt = $pdo->query("SELECT * FROM T_Supplier ORDER BY Denomination");
-    return $stmt->fetchAll(PDO::FETCH_ASSOC);
-}
-
-// Função para buscar todos os protótipos
-function getPrototypes($pdo) {
-    $stmt = $pdo->query("SELECT * FROM T_Prototype ORDER BY Name, Version");
-    return $stmt->fetchAll(PDO::FETCH_ASSOC);
-}
-
-// Função para buscar todos os componentes
-function getComponents($pdo) {
-    $stmt = $pdo->query("
-        SELECT c.*, 
-               m.Denomination as Manufacturer_Name,
-               s.Denomination as Supplier_Name
-        FROM T_Component c
-        LEFT JOIN T_Manufacturer m ON c.Manufacturer_ID = m.Manufacturer_ID
-        LEFT JOIN T_Supplier s ON c.Supplier_ID = s.Supplier_ID
-        ORDER BY c.Denomination
-    ");
-    return $stmt->fetchAll(PDO::FETCH_ASSOC);
-}
 
 // Processar ações CRUD baseadas no entity e action
 switch ($entity) {
