+++ tabs/bomlist/getters.php
@@ -1,0 +1,33 @@
+<?php
+// Função para buscar todos os fabricantes
+function getManufacturers($pdo) {
+    $stmt = $pdo->query("SELECT * FROM T_Manufacturer ORDER BY Denomination");
+    return $stmt->fetchAll(PDO::FETCH_ASSOC);
+}
+
+// Função para buscar todos os fornecedores
+function getSuppliers($pdo) {
+    $stmt = $pdo->query("SELECT * FROM T_Supplier ORDER BY Denomination");
+    return $stmt->fetchAll(PDO::FETCH_ASSOC);
+}
+
+// Função para buscar todos os protótipos
+function getPrototypes($pdo) {
+    $stmt = $pdo->query("SELECT * FROM T_Prototype ORDER BY Name, Version");
+    return $stmt->fetchAll(PDO::FETCH_ASSOC);
+}
+
+// Função para buscar todos os componentes
+function getComponents($pdo) {
+    $stmt = $pdo->query("
+        SELECT c.*, 
+               m.Denomination as Manufacturer_Name,
+               s.Denomination as Supplier_Name
+        FROM T_Component c
+        LEFT JOIN T_Manufacturer m ON c.Manufacturer_ID = m.Manufacturer_ID
+        LEFT JOIN T_Supplier s ON c.Supplier_ID = s.Supplier_ID
+        ORDER BY c.Denomination
+    ");
+    return $stmt->fetchAll(PDO::FETCH_ASSOC);
+}
+?>
